remove: Remove unncessary implementation
