+++ flutter_secure_storage/android/src/main/AndroidManifest.xml
@@ -1,5 +1,2 @@
-<manifest
-    xmlns:tools="http://schemas.android.com/tools"
-    package="com.it_nomads.fluttersecurestorage">
-    <uses-sdk tools:overrideLibrary="androidx.security"/>
+<manifest package="com.it_nomads.fluttersecurestorage">
 </manifest>
